Working GPTree with basic arithmetic operations. Wrote function to output to DOT file. See test/gp/gptrees/GPTreeTest.
Struggling to decide if I should use integers or doubles.
Doubles would be more fine-grained, but since this will be applied to
chess performance is critical, and int's are faster.

Divide seems to be a poor idea with ints, since it will return 0 half
the time. I just had a brilliant idea, though! Do float division then
round back to an integer!!!